# docs/swagger.yaml
definitions:
  dtos.AccessLevel:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - READER
    - USER
    - ADMIN
...
  dtos.PunqUser:
    properties:
      accessLevel:
        $ref: '#/definitions/dtos.AccessLevel'
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
    required:
    - accessLevel
    - createdAt
    - displayName
    - email
    - id
    - password
    type: object
...
info:
  contact: {}
paths:
...
  /user:
    patch:
      parameters:
      - description: PunqUser
        in: body
        name: body
        schema:
          $ref: '#/definitions/dtos.PunqUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PunqUser'
      security:
      - Bearer: []
      tags:
      - User
    post:
      parameters:
      - description: PunqUser
        in: body
        name: body
        schema:
          $ref: '#/definitions/dtos.PunqUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PunqUser'
      security:
      - Bearer: []
      tags:
      - User
  /user/{id}:
    delete:
      parameters:
      - description: ID of the user
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PunqUser'
      security:
      - Bearer: []
      tags:
      - User
    get:
      parameters:
      - description: ID of the user
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PunqUser'
      security:
      - Bearer: []
      tags:
      - User
  /user/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.PunqUser'
            type: array
      security:
      - Bearer: []
      tags:
      - User
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token. Use the user/login
      route to get the token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"